Add script to manage dotfiles setup
The new shell script, setup_dotfiles.sh, automates the installation and updating of dotfiles in the user's home directory. It checks if the dotfiles directory is a symlink, if not, it backs up the existing file if it exists and creates a new symlink. In case the directory already exists, it updates the dotfiles from the Git repository. This script streamlines and simplifies the management of the dotfiles.