Add execution of diyfetch to .zshrc and create diyfetch script
A new script called 'diyfetch' was created. This script, inspired by Neofetch, formats and displays system information when a terminal is opened. To execute the script when a new terminal session is started, the execution command was added to the .zshrc file.

This provides a quick overview of system stats such as operating system, cpu name, shell, date, uptime, and package manager, adding a fun tweak to the terminal.